+++ lib/src/routers/workflow_router.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 from db.db_engine import get_db
 from db.models.workflow_templates import WorkflowTemplates
 from db.models.users import Users
-from blocks import WorkflowTemplate
+from engine.blocks import WorkflowTemplate
 from core.container_utils.docker_tools import DockerTools
 from routers.dependency import get_current_user
