+++ lib/src/server.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI, Request, UploadFile, File, HTTPException
 from typing import Optional
 from fastapi.middleware.cors import CORSMiddleware
 from workflow import RunWorkflow
-from blocks import WorkflowTemplate
+from engine.blocks import WorkflowTemplate
 import json
 
 # Read the escaped JSON from file
