+++ lib/src/workflow.py
@@ -1,6 +1,6 @@
 from collections import defaultdict, deque
 from typing import Any
-from blocks import WorkflowTemplate
+from engine.blocks import WorkflowTemplate
 from implementations.tasks.implementer import Implementer
 from implementations.integrations.implementer import IntegrationImplementer
